Work on some portability problems.
git-svn-id: http://dminfo:8686/svn/repository/models/trunk@565 cdc4813b-4270-ec4f-936e-925f93a782c2

+++ input_storage/src/ItemInputState.cpp
@@ -5,6 +5,7 @@
 #include<sstream>
 #include "FileInputState.h"
 #include "HDFTableManager.h"
+#include "ItemInputState.h"
 
 using namespace boost::algorithm;
 using namespace boost;
@@ -61,7 +62,7 @@ bool ItemInputState<T>::verifyHeader(string& line)
   StringVec colNames = splitLine;   //todo:   get from T
   bool headersOK = splitLine.size() == colNames.size();
   if (! headersOK) return false;
-  for (int i = 0; i < colNames.size() ; ++i)
+  for (size_t i = 0; i < colNames.size() ; ++i)
     {
       headersOK &= (splitLine[i] == colNames[i]);
       if (!headersOK) break;
